% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_darts_to_complete_leg_dataset.R
\name{create_darts_to_complete_leg_dataset}
\alias{create_darts_to_complete_leg_dataset}
\title{Transform Dataset To Calculate Numer of Darts to Win a Leg}
\usage{
create_darts_to_complete_leg_dataset()
}
\value{
data.frame showing the number of darts required to win each of the legs that a player won
}
\description{
Transform Dataset To Calculate Numer of Darts to Win a Leg
}
\examples{
create_darts_to_complete_leg_dataset()
}
