% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_double_success_per_match_dataset.R
\name{create_double_success_per_match_dataset}
\alias{create_double_success_per_match_dataset}
\title{Transform Dataset To Calculate Conversion Percentage of Leg Winning Opportunities}
\usage{
create_double_success_per_match_dataset()
}
\value{
data.frame showing the conversion percentage of each player on match winning shots in each match
}
\description{
Transform Dataset To Calculate Conversion Percentage of Leg Winning Opportunities
}
\examples{
create_double_success_per_match_dataset()
}
