% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pressure_dataset.R
\name{create_pressure_effect_dataset}
\alias{create_pressure_effect_dataset}
\title{Create Pressure Dataset}
\usage{
create_pressure_effect_dataset()
}
\value{
data.frame showing the average conversion of winning opportunities when their shot is under pressure or not
}
\description{
Create Pressure Dataset
}
\examples{
create_pressure_effect_dataset()
}
