% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_to_global_envir.R
\name{load_data_to_global_envir}
\alias{load_data_to_global_envir}
\title{Assign packaged datasets to global environment}
\usage{
load_data_to_global_envir()
}
\value{
Four tibbles containing raw datasets for throws, legs, matches and players
}
\description{
Assign packaged datasets to global environment
}
\examples{
load_data_to_global_envir()
}
